Added monster_def for def file loading. I think I did this in a very good way, I took a lot of inspiration from Quake and other games and professional ways of doing things. I am aiming to make things as professional as I can as sort of practice and use best practices. I've done a bit of research on what games do and whats good and what isnt for file ops as well as some other stuff.

+++ include/projectile.h
@@ -19,7 +19,7 @@ typedef struct {
 	// === TIME ===
 
 	Uint8			spawnImmunity;		//a short window of spawn immunity of the projectile protecting it from being freed on map collision
-	Uint8			explosionTime;		//how long the projectile explodes for
+	Uint32			explosionTime;		//how long the projectile explodes for
 	Uint32			timeAtSpawn;		//the time when the projectile was spawned
 	Uint32			timeAtExplosion;	//the time when the projectile exploded
